2.2. Building our board
Create a small saveStatePlugin so we can listen for mutations being made to our Vuex store, and store them in the local browser storage. This way if we refresh our browser, our board won’t get reset.
We’ll then need to use this plugin and restore the board from local storage if it already exists.