6.6 Dragging columns
Unify the ui. Trigger moveTaskOrColumn at both drop listeners.
The behaviour is now like so:
if we drop a column over a task columns are switched
if we drop a column over a column columns are switched
if we drop a task over a task task dropped is inserted at that place where is dropped
if we drop a task over a column task is placed at the end of the column

The last behaviour is emulated by this code
moveTaskOrColumn(e, toTasks, toColumnIndex, toTaskIndex){
const type = e.dataTransfer.getData('type')

  if ( type === 'task' ){
    this.moveTask(e, toTasks, toTaskIndex !== 'undefined' ? toTaskIndex : toTasks.length)
  } else {
    this.moveColumn(e, toColumnIndex)
  }
}

If task is droped over a task toTaskIndex is defined, but if task is dropped over a column we don't have any taskIndex so we place that task at the end of the column (toTasks.length)